Struct strava::athletes::Athlete

pub struct Athlete {
    pub id: i32,
    pub resource_state: ResourceState,
    pub firstname: Option<String>,
    pub lastname: Option<String>,
    pub profile_medium: Option<String>,
    pub profile: Option<String>,
    pub city: Option<String>,
    pub state: Option<String>,
    pub country: Option<String>,
    pub sex: Option<String>,
    pub friend: Option<String>,
    pub follower: Option<String>,
    pub premium: Option<bool>,
    pub created_at: Option<String>,
    pub updated_at: Option<String>,
    pub approve_followers: Option<bool>,
    pub follower_count: Option<i32>,
    pub friend_count: Option<i32>,
    pub mutual_friend_count: Option<i32>,
    pub date_preference: Option<String>,
    pub measurement_preference: Option<String>,
    pub email: Option<String>,
    pub ftp: Option<i32>,
    pub weight: Option<f32>,
}

A strava athlete

The object may be returned in detailed, summary or meta representations.

See: http://strava.github.io/api/v3/athlete/

Fields

id

Athlete's ID on strava

resource_state

The resource state dictates which fields should be safe to unwrap

firstname

First name

lastname

Last name

profile_medium

URL to a 62x62 pixel profile picture

profile

URL to a 124x124 pixel profile picture

city

home city

state

home state

country

home country

sex

gender

friend

"pending", "accepted", "blocked" or None - the authenticated athlete’s following status of this athlete

follower

"pending", "accepted", "blocked" or None - this athlete's following status of the authenticated athlete

premium

The athlete uses strava's premium features

created_at

updated_at

approve_followers

follower_count

friend_count

mutual_friend_count

date_preference

measurement_preference

feet or meters

email

ftp

weight

Methods

impl Athlete

fn get_current(token: &AccessToken) -> Result<Athlete>

Get the athlete associated with the given access token

fn get(token: &AccessToken, id: i32) -> Result<Athlete>

Get an Athlete by id

fn stats(&self, token: &AccessToken) -> Result<Stats>

Get stats for an athlete.

This is only available for the currently authenticated athlete

fn koms(&self, token: &AccessToken) -> Result<Paginated<SegmentEffort>>

Get all KOMs for the Athlete.